Data table working
- Data table updates values when data is received.
- PSPCOM acceleration, gyro, temp, and pressure are supported currently.
- Fixed some issues in the PSPCOM CRC and state machine.

+++ pspcom/pspcommsg.h
@@ -5,7 +5,7 @@
 
 typedef struct {
     uint8_t payload_len;
-    uint16_t device_id;
+    uint8_t device_id;
     uint8_t msg_id;
     uint8_t payload[256];
 } pspcommsg;
